change compile problems

+++ src/tcpcopy/send.h
@@ -5,7 +5,7 @@ int send_init();
 int send_close();
 uint32_t send_ip_packet(uint64_t fake_ip_addr,
 		unsigned char *data,uint32_t ack_seq,uint32_t* sendSeq,
-		uint32_t* sendConPackets);
+		size_t* sendConPackets);
 
 
 #endif   /* ----- #ifndef _TCP_REDIRECT_SEND_H_INC  ----- */
